requirements.txt
Django==3.2.6
djangorestframework==3.12.4
django-filter==2.4.0
python-dotenv==0.19.0
django-model-utils==3.1.2
